Conflicts:
	serialisation/include/serialisation/arrayInfo.h
	serialisation/include/serialisation/binaryDeserMessage.h
	serialisation/include/serialisation/interface/abstractTempArray.h

+++ serialisation/include/serialisation/binarySerMessage.h
@@ -236,7 +236,7 @@ inline void BinarySerMessage::StoreVector( std::vector< double > &container, uin
             intBuffer[ k ] = Util::DoubleToUInt64( firstFloat[ k ] );
         }
 
-        mStreamBuffer.WriteBytes( intBuffer, blockSize * sizeof( uint32_t ) );
+        mStreamBuffer.WriteBytes( intBuffer, blockSize * sizeof( uint64_t ) );
     }
 }
